backtest-engine-paper/requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
matplotlib>=3.7
tqdm>=4.66
pytest>=8.0
python-dateutil>=2.8
